+++ src/main/java/io/github/artlibs/autotrace4j/support/ClassUtils.java
@@ -1,6 +1,6 @@
 package io.github.artlibs.autotrace4j.support;
 
-import io.github.artlibs.autotrace4j.support.exception.WalkClassFileException;
+import io.github.artlibs.autotrace4j.exception.WalkClassFileException;
 import net.bytebuddy.dynamic.ClassFileLocator;
 import net.bytebuddy.dynamic.loading.ClassInjector;
